refactoring the existing CompositionalConfig class

+++ src/opm/input/eclipse/EclipseState/Runspec.cpp
@@ -805,7 +805,7 @@ bool Runspec::compostionalMode() const
     return this->m_comps > 0;
 }
 
-int Runspec::numComps() const
+size_t Runspec::numComps() const
 {
     return this->m_comps;
 }
